+++ tools/disaster-recovery/readiness-evidence-schema.ts
@@ -1,0 +1,337 @@
+import { canonicalJson } from './canonical-json.ts'
+import {
+	type EvidenceDetailsByKind,
+	type EvidenceKind,
+	type SignedEvidenceEnvelope,
+	canonicalContracts,
+	maximumSupportedD1BackupBytes,
+} from './readiness-contracts.ts'
+import {
+	base64Pattern,
+	exactKeys,
+	isIsoDate,
+	isNonemptyString,
+	isNonnegativeInteger,
+	isPositiveInteger,
+	isRecord,
+	parseIdentity,
+	sha256Pattern,
+	uuidPattern,
+} from './readiness-validation.ts'
+
+const resourceIds = new Set<string>(
+	canonicalContracts.map((contract) => contract.id),
+)
+const evidenceKinds = new Set<string>(
+	canonicalContracts.flatMap((contract) => contract.requiredEvidenceKinds),
+)
+
+function parseDetails<K extends EvidenceKind>(
+	kind: K,
+	value: unknown,
+): EvidenceDetailsByKind[K] | undefined {
+	if (!isRecord(value)) return undefined
+	switch (kind) {
+		case 'alarm-rebuild-drill':
+			if (
+				exactKeys(value, ['alarmsRebuilt', 'deliveryVerified']) &&
+				isNonnegativeInteger(value.alarmsRebuilt) &&
+				value.deliveryVerified === true
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'artifact-mirror-verification':
+			if (
+				exactKeys(value, ['mirrorSha256', 'refCount']) &&
+				typeof value.mirrorSha256 === 'string' &&
+				sha256Pattern.test(value.mirrorSha256) &&
+				isNonnegativeInteger(value.refCount)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'bundle-kv-rebuild-drill':
+			if (
+				exactKeys(value, ['contentSha256', 'keyCount']) &&
+				typeof value.contentSha256 === 'string' &&
+				sha256Pattern.test(value.contentSha256) &&
+				isNonnegativeInteger(value.keyCount)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'community-icon-rebuild-drill':
+			if (
+				exactKeys(value, ['contentSha256', 'objectCount']) &&
+				typeof value.contentSha256 === 'string' &&
+				sha256Pattern.test(value.contentSha256) &&
+				isNonnegativeInteger(value.objectCount)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'contract-verification':
+			if (
+				exactKeys(value, ['checksPassed', 'contractVersion']) &&
+				isPositiveInteger(value.checksPassed) &&
+				isNonemptyString(value.contractVersion)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'destination-credential-check':
+		case 'source-credential-check':
+			if (
+				exactKeys(value, ['credentialId', 'scope']) &&
+				isNonemptyString(value.credentialId) &&
+				isNonemptyString(value.scope)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'd1-restore-drill':
+			if (
+				exactKeys(value, [
+					'foreignKeyViolations',
+					'quickCheck',
+					'restoredDatabaseUuid',
+				]) &&
+				value.foreignKeyViolations === 0 &&
+				value.quickCheck === 'ok' &&
+				typeof value.restoredDatabaseUuid === 'string' &&
+				uuidPattern.test(value.restoredDatabaseUuid)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'd1-size-ceiling-check':
+			if (
+				exactKeys(value, [
+					'ceilingBytes',
+					'measuredBytes',
+					'monitoredAt',
+					'sourceAccountId',
+					'sourceDatabaseUuid',
+				]) &&
+				isPositiveInteger(value.ceilingBytes) &&
+				isNonnegativeInteger(value.measuredBytes) &&
+				value.ceilingBytes <= maximumSupportedD1BackupBytes &&
+				value.measuredBytes < value.ceilingBytes &&
+				isIsoDate(value.monitoredAt) &&
+				isNonemptyString(value.sourceAccountId) &&
+				typeof value.sourceDatabaseUuid === 'string' &&
+				uuidPattern.test(value.sourceDatabaseUuid)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'escrow-recovery-drill':
+			if (
+				exactKeys(value, ['custodian', 'recoveredFingerprint']) &&
+				isNonemptyString(value.custodian) &&
+				isNonemptyString(value.recoveredFingerprint)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'inventory':
+			if (
+				exactKeys(value, ['inventorySha256', 'itemCount']) &&
+				typeof value.inventorySha256 === 'string' &&
+				sha256Pattern.test(value.inventorySha256) &&
+				isNonnegativeInteger(value.itemCount)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'key-fingerprint':
+			if (
+				exactKeys(value, [
+					'destinationFingerprint',
+					'matched',
+					'sourceFingerprint',
+				]) &&
+				isNonemptyString(value.destinationFingerprint) &&
+				value.matched === true &&
+				isNonemptyString(value.sourceFingerprint) &&
+				value.destinationFingerprint === value.sourceFingerprint
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'oauth-reauthorization-drill':
+			if (
+				exactKeys(value, ['connectorCount', 'reauthorizedCount']) &&
+				isNonnegativeInteger(value.connectorCount) &&
+				isNonnegativeInteger(value.reauthorizedCount) &&
+				value.reauthorizedCount === value.connectorCount
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'queue-workflow-rebuild-drill':
+			if (
+				exactKeys(value, ['deliveryVerified', 'queueCount', 'workflowCount']) &&
+				value.deliveryVerified === true &&
+				isNonnegativeInteger(value.queueCount) &&
+				isNonnegativeInteger(value.workflowCount)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'r2-round-trip-drill':
+			if (
+				exactKeys(value, ['bytes', 'objectKey', 'sha256']) &&
+				isNonnegativeInteger(value.bytes) &&
+				isNonemptyString(value.objectKey) &&
+				typeof value.sha256 === 'string' &&
+				sha256Pattern.test(value.sha256)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'storage-runner-round-trip-drill':
+			if (
+				exactKeys(value, ['instanceCount', 'kvEntries', 'sqliteRows']) &&
+				isNonnegativeInteger(value.instanceCount) &&
+				isNonnegativeInteger(value.kvEntries) &&
+				isNonnegativeInteger(value.sqliteRows)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'transfer-support-check':
+			if (
+				exactKeys(value, ['mechanism', 'supported']) &&
+				isNonemptyString(value.mechanism) &&
+				value.supported === true
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		case 'vectorize-rebuild-drill':
+			if (
+				exactKeys(value, ['queryVerified', 'vectorCount']) &&
+				value.queryVerified === true &&
+				isNonnegativeInteger(value.vectorCount)
+			) {
+				return value as EvidenceDetailsByKind[K]
+			}
+			return undefined
+		default: {
+			const exhaustive: never = kind
+			throw new Error(String(exhaustive))
+		}
+	}
+}
+
+function destinationIsRequired(kind: EvidenceKind): boolean {
+	switch (kind) {
+		case 'inventory':
+		case 'source-credential-check':
+		case 'd1-size-ceiling-check':
+			return false
+		case 'alarm-rebuild-drill':
+		case 'artifact-mirror-verification':
+		case 'bundle-kv-rebuild-drill':
+		case 'community-icon-rebuild-drill':
+		case 'contract-verification':
+		case 'destination-credential-check':
+		case 'd1-restore-drill':
+		case 'escrow-recovery-drill':
+		case 'key-fingerprint':
+		case 'oauth-reauthorization-drill':
+		case 'queue-workflow-rebuild-drill':
+		case 'r2-round-trip-drill':
+		case 'storage-runner-round-trip-drill':
+		case 'transfer-support-check':
+		case 'vectorize-rebuild-drill':
+			return true
+		default: {
+			const exhaustive: never = kind
+			throw new Error(String(exhaustive))
+		}
+	}
+}
+
+export function parseSignedEvidenceEnvelope(
+	input: unknown,
+): SignedEvidenceEnvelope | undefined {
+	if (
+		!isRecord(input) ||
+		!exactKeys(input, ['content', 'schemaVersion', 'signature']) ||
+		input.schemaVersion !== 1 ||
+		!isRecord(input.content) ||
+		!exactKeys(input.content, [
+			'changeId',
+			'destinationIdentity',
+			'details',
+			'kind',
+			'outcome',
+			'performedAt',
+			'resourceId',
+			'sourceIdentity',
+			'systemVersion',
+			'uri',
+			'verifierIdentity',
+		]) ||
+		typeof input.content.kind !== 'string' ||
+		!evidenceKinds.has(input.content.kind) ||
+		typeof input.content.resourceId !== 'string' ||
+		!resourceIds.has(input.content.resourceId) ||
+		input.content.outcome !== 'passed' ||
+		!isNonemptyString(input.content.uri) ||
+		!isNonemptyString(input.content.verifierIdentity) ||
+		!isNonemptyString(input.content.changeId) ||
+		!isNonemptyString(input.content.systemVersion) ||
+		!isIsoDate(input.content.performedAt)
+	) {
+		return undefined
+	}
+	const kind = input.content.kind as EvidenceKind
+	const sourceIdentity = parseIdentity(input.content.sourceIdentity)
+	const destinationIdentity =
+		input.content.destinationIdentity === null
+			? null
+			: parseIdentity(input.content.destinationIdentity)
+	const details = parseDetails(kind, input.content.details)
+	if (
+		!sourceIdentity ||
+		destinationIdentity === undefined ||
+		(destinationIsRequired(kind)
+			? destinationIdentity === null
+			: destinationIdentity !== null) ||
+		!details ||
+		!isRecord(input.signature) ||
+		!exactKeys(input.signature, ['algorithm', 'keyId', 'value']) ||
+		input.signature.algorithm !== 'Ed25519' ||
+		!isNonemptyString(input.signature.keyId) ||
+		typeof input.signature.value !== 'string' ||
+		!base64Pattern.test(input.signature.value) ||
+		Buffer.from(input.signature.value, 'base64').byteLength !== 64
+	) {
+		return undefined
+	}
+	if (kind === 'd1-size-ceiling-check') {
+		const sizeDetails =
+			details as EvidenceDetailsByKind['d1-size-ceiling-check']
+		if (
+			sizeDetails.sourceAccountId !== sourceIdentity.accountId ||
+			sizeDetails.sourceDatabaseUuid !== sourceIdentity.resourceId ||
+			sizeDetails.monitoredAt !== input.content.performedAt
+		) {
+			return undefined
+		}
+	}
+	return input as SignedEvidenceEnvelope
+}
+
+export function canonicalEvidencePayload(
+	envelope: Pick<SignedEvidenceEnvelope, 'schemaVersion' | 'content'>,
+): string {
+	return canonicalJson({
+		schemaVersion: envelope.schemaVersion,
+		content: envelope.content,
+	})
+}
